+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fixed `Markdown.goto_anchor` no longer scrolling the heading into view https://github.com/Textualize/textual/pull/4583
 - Fixed Footer flicker on initial focus https://github.com/Textualize/textual/issues/4573
 
+### Added
+
+- Added templates support to `Input`
+
 ## [0.63.6] - 2024-05-29
 
 ### Fixed